change the main file
the library is intended to be used with a bundler, even though for some reason webpack still bundles the bundled p5.min.js, I thought that you should use this library as a module, especially if you are working in source code editors like VS Code. I don't know why this library had a main of a bundled version that literally is to be used in a global context(a script tag etc.). I noticed this because every other library like socket.io-client has a main of the index.js file, that isn't bundled.